Fix timeout error on updating sources and languages (#42)
This PR fixes a timeout issue when updating sources and languages on slow connections.  It raises the TTL from 5s to 30s.  

The TTL is set in the underlying library unfoldingword.tools.http.  Unfortunately the TTL is not configurable within the library, so for now I have replaced the dependency with a slightly modified version of the library.

- Remove existing library unfoldingword.tools.http:2.4.2 from dependencies list.
- Exclude indirect dependencies from other libraries.
- Add a copy of unfoldingword.tools.http:2.4.2 to this source tree.
- Modify our copy to increase the TTL from 5s to 30s.
- Update version number and version code.