Replace Account Creation form with link (#43)
This PR changes the behavior of the "Create New Server Account" button in the Profile Activity.  Instead of opening the create user form, it will now open the user's default browser to the server account creation page.  This page is configurable, and comes with default values for WACS and DCS.

This is necessary because of changes to how Gitea manages the account creation process.  The way the BW previously created accounts now results in a confusing message and a frustrating workflow.  This change now directs the user to the server to create their account, which they can then use to log in to BTT-Writer.